debugging combining images of different sizes and pixel sizes

+++ src/jaz/obs_model.h
@@ -56,7 +56,7 @@ class ObservationModel
 
 
 	protected:
-			
+
 			// cached values - protected to prevent users from accidentally changing them,
 			// expecting the changes to propagate into the optics star-file
 			std::vector<double> angpix, lambda, Cs;
@@ -103,14 +103,14 @@ class ObservationModel
 
 		// effect of symmetric aberration (cached)
 		const Image<RFLOAT>& getGammaOffset(int optGroup, int s);
-		
+
 		Matrix2D<RFLOAT> applyAnisoMagTransp(
 				Matrix2D<RFLOAT> A3D_transp, int opticsGroup);
-		
+
 		Matrix2D<RFLOAT> getMag3x3(int opticsGroup);
-		
+
 		Matrix2D<RFLOAT> applyScaleDifference(
-				Matrix2D<RFLOAT> A3D_transp, int opticsGroup, 
+				Matrix2D<RFLOAT> A3D_transp, int opticsGroup,
 				int s3D, double angpix3D);
 
 
@@ -122,23 +122,23 @@ class ObservationModel
 
         double angToPix(double a, int s, int opticsGroup = 0) const;
         double pixToAng(double p, int s, int opticsGroup = 0) const;
-		
+
 		double getPixelSize(int opticsGroup) const;
 		std::vector<double> getPixelSizes() const;
-		
+
 		double getWavelength(int opticsGroup) const;
 		std::vector<double> getWavelengths() const;
-		
+
 		double getSphericalAberration(int opticsGroup) const;
 		std::vector<double> getSphericalAberrations() const;
-	
+
 		int getBoxSize(int opticsGroup) const;
 		void getBoxSizes(std::vector<int>& sDest, std::vector<int>& shDest) const;
 
 		Matrix2D<RFLOAT> getMagMatrix(int opticsGroup) const;
 		std::vector<Matrix2D<RFLOAT> > getMagMatrices() const;
-		
-		int getOpticsGroup(const MetaDataTable &particlesMdt, int particle) const;
+
+		int getOpticsGroup(const MetaDataTable &particlesMdt, long int particle) const;
 
 		/* duh */
 		int numberOfOpticsGroups() const;
@@ -158,10 +158,10 @@ class ObservationModel
 		   and translate the indices in particle table partMdt.
 		   (Merely changing the order in opticsMdt would fail if groups were missing.) */
 		void sortOpticsGroups(MetaDataTable& partMdt);
-		
+
 		/* Return the set of optics groups present in partMdt */
 		std::vector<int> getOptGroupsPresent_oneBased(const MetaDataTable& partMdt) const;
-		
+
 		/* Return the set of optics groups present in partMdt */
 		std::vector<int> getOptGroupsPresent_zeroBased(const MetaDataTable& partMdt) const;
